requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Configuration and validation
pyyaml==6.0.1
jsonschema==4.20.0

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# HTTP and async
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Database clients
redis==5.0.1
asyncpg==0.29.0
neo4j==5.15.0
qdrant-client==1.7.0
sqlalchemy==2.0.23

# AI/ML
openai==1.6.1
anthropic==0.8.1
transformers==4.36.2
sentence-transformers==2.2.2
torch==2.1.2
numpy==1.26.2
pandas==2.1.4

# Utilities
click==8.1.7
rich==13.7.0
tenacity==8.2.3
cachetools==5.3.2

# Data processing
pypdf==3.17.4
beautifulsoup4==4.12.2
lxml==4.9.4

# Testing helpers (also needed in production for health checks)
pytest-asyncio==0.21.1